refactor: reorganize discovery
refactor: more moves


refactor: more reorg
